// pseudoWeb/packages/core/src/lexer/types/tokenType.ts
// Enumeración de tipos de tokens para el intérprete de pseudocódigo

export enum TokenType {
  Accion = "Accion",
  FinAccion = "FinAccion",
  Si = "Si",
  SiNo = "SiNo",
  FinSi = "FinSi",
  Mientras = "Mientras",
  FinMientras = "FinMientras",
  Para = "Para",
  FinPara = "FinPara",
  Repetir = "Repetir",
  Hasta = "Hasta",
  HastaQue = "HastaQue",
  Segun = "Segun",
  Hacer = "Hacer",
  Otro = "Otro",
  FinSegun = "FinSegun",
  FinFuncion = "FinFuncion",
  FinProcedimiento = "FinProcedimiento",
  Leer = "Leer",
  Escribir = "Escribir",
  Funcion = "Funcion",
  Procedimiento = "Procedimiento",
  Entonces = "Entonces",
  ES = "ES",
  Ambiente = "Ambiente",
  Proceso = "Proceso",
  Verdadero = "Verdadero",
  Falso = "Falso",

  Asignacion = "Asignacion",
  Suma = "Suma",
  Resta = "Resta",
  Multiplicacion = "Multiplicacion",
  Division = "Division",
  Div = "Div",
  Mod = "Mod",
  Potencia = "Potencia",

  Igual = "Igual",
  Distinto = "Distinto",
  Menor = "Menor",
  Mayor = "Mayor",
  MenorIgual = "MenorIgual",
  MayorIgual = "MayorIgual",

  Y = "Y",
  O = "O",
  No = "No",

  ParentesisIzquierdo = "ParentesisIzquierdo",
  ParentesisDerecho = "ParentesisDerecho",
  DosPuntos = "DosPuntos",
  Coma = "Coma",
  Rango = "Rango",

  Entero = "Entero",
  Real = "Real",
  Caracter = "Caracter",
  Logico = "Logico",
  Alfanumerico = "Alfanumerico",

  Identificador = "Identificador",
  Comentario = "Comentario",
  Desconocido = "Desconocido",
  EOF = "EOF",
  PuntoYComa = "PuntoYComa",
  SaltoDeLinea = "SaltoDeLinea",
}